backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.2.35
cryptography==40.0.2
daphne==3.0.2
Django==4.2
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-model-utils==4.3.1
django-push-notifications==3.0.0
django-rest-framework==0.1.0
django-snitch==3.1.7
django-timezone-field==5.0
django-websocket-notifications==1.2.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.5
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
service-identity==21.1.0
single-source==0.3.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
zope.interface==6.0
